+++ mapInfo.json
@@ -1,0 +1,8 @@
+[
+    {
+        "title": "test",
+        "description": "test",
+        "max": 999,
+        "min": 1
+    }
+]
