--- test.js
@@ -1,6 +1,0 @@
-ob = {
-  apple: "ddd",
-  orange: "ww",
-};
-
-console.log(ob.apple);
